strategy/supertrend: refactor to smaller functions
